requirements/dev.txt
-r base.txt
pytest
pytest-django
django-debug-toolbar
